Major League Baseball Teams Quiz (TermProject.py)
Video Demo: https://youtu.be/ltYp0A3koPs?si=A8TE9kOMjFrNb4aN

Description: This program tests the knowledge of sports enthusiasts about Major League baseball teams and the cities they are associated with. It is meant for entertainment and could be used as a tool to test memory.

Key elements of the program include an "import random" module. The random module is used to choose an unpredictable team and state to quiz the user on team and state affiliation. This allows for a simulation that is similar to shuffling cards in a deck. A dictionary is built using a mutable data structure that stores the name of a Major League Baseball (MLB) team and the city to which it belongs. The key-value string pairs are in the format "name : city, state". There are four functions defined, starting with a function that counts the number of MLB teams and city pairs, and requiring a number to be entered. num_questions = questions_count(len(teams)) is the function that returns an integer. Uses a ”while True” loop to prompt the user for at least one input to start the quiz. “except ValueError:” returns an error if the user does not enter a number and prevents the program from crashing. The next function takes in three inputs: team, correct city, and all the cities. “wrong_choices = random.sample” builds a list of three random wrong answers out of the four. Leaves one city and team as the correct answer. “options = wrong_choices + [correct_city]” adds the correct city to the list of wrong cities. Displays the list of options. Uses a ”while True:” loop until a valid number is entered. “return options[choice – 1] == correct_city” checks the correctness of the answer. Takes in the parameters that list all the MLB teams (teams), lists the cities (cities), creates a dictionary mapping teams to cities (team_city_map), and takes in how many questions (num_questions). Zeros out the score by starting it at 0 (score = 0) and stores the teams the user answers correctly (missed = []). Randomly shuffles the teams (random.shuffle(teams)) and creates a quiz loop running once per question and stops at the user-selected number “(for i in range(num_questions):”. Gets the current team and correct city using team = teams[i] and correct_city = team_city_map[team]. Asks the question using an “if /else” statement, updates the score, and returns the results. Takes inputs for number of correct answers (score), questions asked (num_questions), questions answered incorrectly (missed), and the dictionary of teams (team_city_map). The program assigns a final score based on the number of choices made (num_score) and the number of correct answers (score). It also provides a percentage accuracy by dividing the number of correct answers by the total number of questions and multiplying by 100. If anything was missed, it is printed in the “for” statement (for team in missed:).